More type and include cleanups.

+++ Source/KSCrash-Tests/KSFileUtils_Tests.m
@@ -135,10 +135,10 @@
     int fd = open([path UTF8String], O_RDONLY);
     XCTAssertTrue(fd >= 0, @"");
     char* bytes;
-    size_t readLength;
+    int readLength;
     bool result = ksfu_readEntireFile([path UTF8String], &bytes, &readLength);
     XCTAssertTrue(result, @"");
-    NSMutableData* data = [NSMutableData dataWithBytesNoCopy:bytes length:readLength freeWhenDone:YES];
+    NSMutableData* data = [NSMutableData dataWithBytesNoCopy:bytes length:(unsigned)readLength freeWhenDone:YES];
     NSString* actual = [[NSString alloc] initWithData:data encoding:NSUTF8StringEncoding];
     XCTAssertNil(error, @"");
     XCTAssertEqualObjects(actual, expected, @"");
@@ -161,10 +161,10 @@
     int fd = open([path UTF8String], O_RDONLY);
     XCTAssertTrue(fd >= 0, @"");
     char* bytes;
-    size_t readLength;
+    int readLength;
     bool result = ksfu_readEntireFile([path UTF8String], &bytes, &readLength);
     XCTAssertTrue(result, @"");
-    NSMutableData* actual = [NSMutableData dataWithBytesNoCopy:bytes length:readLength freeWhenDone:YES];
+    NSMutableData* actual = [NSMutableData dataWithBytesNoCopy:bytes length:(unsigned)readLength freeWhenDone:YES];
     XCTAssertEqualObjects(actual, expected, @"");
 }
 
@@ -236,7 +236,7 @@
     int fd = open([path UTF8String], O_RDONLY);
     XCTAssertTrue(fd >= 0, @"");
     NSMutableData* data = [NSMutableData dataWithLength:100];
-    ssize_t bytesRead;
+    int bytesRead;
     NSString* actual;
 
     bytesRead = ksfu_readLineFromFD(fd, [data mutableBytes], 100);
